@@ -2,45 +2,57 @@
 
     <PropertyGroup>
         <OutputType>WinExe</OutputType>
-        <TargetFramework>net6.0-windows</TargetFramework>
+        <TargetFramework>net10.0-windows</TargetFramework>
         <Nullable>enable</Nullable>
         <UseWPF>true</UseWPF>
-        <AssemblyVersion>1.0.5.2</AssemblyVersion>
-        <FileVersion>1.0.5.2</FileVersion>
-        <ProductVersion>1.0.1</ProductVersion>
-        <SignAssembly>true</SignAssembly>
-        <AssemblyOriginatorKeyFile>wssids.snk</AssemblyOriginatorKeyFile>
-        <PublicSign>true</PublicSign>
         <Company>WssIDs</Company>
         <ApplicationIcon>main.ico</ApplicationIcon>
-        <Version>1.0.5</Version>
         <PackageIcon>main.ico</PackageIcon>
         <Authors>WssIDs</Authors>
-        <Copyright>2023</Copyright>
+        <Copyright>2025</Copyright>
+        <IsPackable>false</IsPackable>
+		<SatelliteResourceLanguages>true</SatelliteResourceLanguages>
+        <NeutralLanguage>ru</NeutralLanguage>
+		<EnableConfigurationBindingGenerator>true</EnableConfigurationBindingGenerator>
     </PropertyGroup>
 
-    <ItemGroup>
-      <None Remove="Images\logo.png" />
-      <PackageReference Include="Hardcodet.NotifyIcon.Wpf" Version="1.1.0" />
-      <PackageReference Include="MahApps.Metro.IconPacks.Material" Version="4.11.0" />
-      <PackageReference Include="MaterialDesignThemes" Version="4.9.0" />
-      <PackageReference Include="MaterialDesignThemes.MahApps" Version="0.3.0" />
-      <PackageReference Include="Meziantou.Framework.Win32.Dialogs" Version="1.1.4" />
-      <PackageReference Include="Microsoft.Extensions.Configuration" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0" />
-      <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.77" />
-      <PackageReference Include="ShowMeTheXAML.AvalonEdit" Version="2.0.0" />
-      <PackageReference Include="ShowMeTheXAML.MSBuild" Version="2.0.0" />
-      <PackageReference Include="YandexDisk.Client" Version="1.3.27" />
+	<PropertyGroup>
+		<!-- Включаем Interceptors -->
+		<InterceptorsNamespaces>
+			$(InterceptorsNamespaces);Microsoft.Extensions.Configuration.Binder.SourceGeneration
+		</InterceptorsNamespaces>
+	</PropertyGroup>
+
+	<PropertyGroup>
+		<FileVersion>1.1.17.40</FileVersion>
+		<AssemblyVersion>1.1.17.40</AssemblyVersion>
+		<ProductVersion>1.1.17.40</ProductVersion>
+		<Version>1.1.17.40</Version>
+	</PropertyGroup>
+
+	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+	  <DebugType>embedded</DebugType>
+	</PropertyGroup>
+
+	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+	  <DebugType>embedded</DebugType>
+	</PropertyGroup>
+
+	<ItemGroup>
+      <PackageReference Include="Hardcodet.NotifyIcon.Wpf" Version="2.0.1" />
+      <PackageReference Include="MahApps.Metro.IconPacks.Modern" Version="6.2.1" />
+      <PackageReference Include="Microsoft.Extensions.Hosting" Version="10.0.1" />
+      <PackageReference Include="Meziantou.Framework.Win32.Dialogs" Version="1.1.10" />
+      <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.135" />
+      <PackageReference Include="ModernWpfUI" Version="0.9.6" />
       <Resource Include="Images\logo.png" />
     </ItemGroup>
 
     <ItemGroup>
-      <None Remove="appsettings.json" />
       <Content Include="appsettings.json">
+        <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+      </Content>
+      <Content Include="readme.md">
         <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
       </Content>
     </ItemGroup>
