@@ -5,16 +5,13 @@ namespace ElvUIDownloader;
 public static class StreamExtensions
 {
     public static async Task CopyToAsync(this Stream source, Stream destination, int bufferSize, IProgress<long>? progress = null, CancellationToken cancellationToken = default) {
-        if (source == null)
-            throw new ArgumentNullException(nameof(source));
+        ArgumentNullException.ThrowIfNull(source);
         if (!source.CanRead)
             throw new ArgumentException("Has to be readable", nameof(source));
-        if (destination == null)
-            throw new ArgumentNullException(nameof(destination));
+        ArgumentNullException.ThrowIfNull(destination);
         if (!destination.CanWrite)
             throw new ArgumentException("Has to be writable", nameof(destination));
-        if (bufferSize < 0)
-            throw new ArgumentOutOfRangeException(nameof(bufferSize));
+        ArgumentOutOfRangeException.ThrowIfNegative(bufferSize);
 
         var buffer = new byte[bufferSize];
         long totalBytesRead = 0;
