@@ -1,21 +1,56 @@
-﻿namespace ElvUIDownloader.Models.Web;
+﻿#nullable disable
+
+using System.Text.Json.Serialization;
+
+namespace ElvUIDownloader.Models.Web;
 
 public class ElvUiModel
 {
-    public int id { get; set; }
-    public string slug { get; set; }
-    public string author { get; set; }
-    public string name { get; set; }
-    public string url { get; set; }
-    public string version { get; set; }
-    public string changelog_url { get; set; }
-    public string ticket_url { get; set; }
-    public string git_url { get; set; }
-    public List<string> patch { get; set; }
-    public string last_update { get; set; }
-    public string web_url { get; set; }
-    public string donate_url { get; set; }
-    public string small_desc { get; set; }
-    public string screenshot_url { get; set; }
-    public List<string> directories { get; set; }
+    [JsonPropertyName("id")]
+    public int Id { get; set; }
+
+    [JsonPropertyName("slug")]
+    public string Slug { get; set; }
+
+    [JsonPropertyName("author")]
+    public string Author { get; set; }
+
+    [JsonPropertyName("name")]
+    public string Name { get; set; }
+
+    [JsonPropertyName("url")]
+    public string Url { get; set; }
+
+    [JsonPropertyName("version")]
+    public string Version { get; set; }
+
+    [JsonPropertyName("changelog_url")]
+    public string ChangelogUrl { get; set; }
+
+    [JsonPropertyName("ticket_url")]
+    public string TicketUrl { get; set; }
+
+    [JsonPropertyName("git_url")]
+    public string GitUrl { get; set; }
+
+    [JsonPropertyName("patch")]
+    public List<string> Patch { get; set; }
+
+    [JsonPropertyName("last_update")]
+    public string LastUpdate { get; set; }
+
+    [JsonPropertyName("web_url")]
+    public string WebUrl { get; set; }
+
+    [JsonPropertyName("donate_url")]
+    public string DonateUrl { get; set; }
+
+    [JsonPropertyName("small_desc")]
+    public string SmallDesc { get; set; }
+
+    [JsonPropertyName("screenshot_url")]
+    public string ScreenshotUrl { get; set; }
+
+    [JsonPropertyName("directories")]
+    public List<string> Directories { get; set; } = [];
 }
